dom/media/webaudio/test/test_nodeCreationDocumentGone.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test whether we can create an AudioContext interface</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<pre id="test">
<script class="testbody" type="text/javascript">

SimpleTest.requestCompleteLog();
SimpleTest.waitForExplicitFinish();

var a = window.open("file_nodeCreationDocumentGone.html");
a.onbeforeunload = function() {
  setTimeout(function(){
    try {
      a.context.createScriptProcessor(512, 1, 1);
    } catch(e) {
      ok (true,"got exception");
    }
    setTimeout(function() {
      ok (true,"no crash");
      SimpleTest.finish();
    }, 0);
  }, 0);
}


</script>
</pre>
</body>
</html>
